refactor(cublas): Invert build tags and split stub/cuda implementations
Inverted build tag logic: stub code now uses `!cuda && !cuda11 && !cuda12 && !cuda13`, cuda code uses `cuda || cuda11 || cuda12 || cuda13`. Renamed `*_stub.go` files to base names, `*.go` files to `*_cuda.go`. Added README with package description. Removed C imports from stub implementations, replaced with hardcoded constants and error returns. Exposed SrotmParams fields in stub version. Simplified type definitions in stub code
