+++ cublas/README.md
@@ -1,0 +1,3 @@
+# cublas
+
+This package provides a Go interface to the cuBLAS library, which is a GPU-accelerated linear algebra library.
