+++ cublas/cgoflags.go
@@ -1,5 +1,4 @@
-//go:build cuda
-// +build cuda
+//go:build cuda || cuda11 || cuda12 || cuda13
 
 package cublas
