+++ cublas/check.go
@@ -1,12 +1,7 @@
-//go:build cuda
-// +build cuda
+//go:build !cuda && !cuda11 && !cuda12 && !cuda13
 
 package cublas
 
-/*
-#include <cublas_v2.h>
-*/
-import "C"
 import "fmt"
 
 // Status is the cublas Check.
@@ -20,25 +15,24 @@ func (err Status) String() string {
 	return fmt.Sprintf("Unknown Error Code:%d", err)
 }
 
-func Check(x C.cublasStatus_t) error {
-	err := Status(x)
-	if err == CublasStatusSuccess {
+func Check(x Status) error {
+	if x == CublasStatusSuccess {
 		return nil
 	}
-	return err
+	return x
 }
 
 const (
-	CublasStatusSuccess        Status = C.CUBLAS_STATUS_SUCCESS          // The operation completed successfully.
-	CublasStatusNotInitialized Status = C.CUBLAS_STATUS_NOT_INITIALIZED  // The cuBLAS library was not initialized. This is usually caused by the lack of a prior cublasCreate() call,
-	CublasStatusAllocFailed    Status = C.CUBLAS_STATUS_ALLOC_FAILED     // Resource allocation failed inside the cuBLAS library.
-	CublasStatusInvalidValue   Status = C.CUBLAS_STATUS_INVALID_VALUE    // An unsupported value or parameter was passed to the function (a negative vector size, for example).
-	CublasStatusArchMismatch   Status = C.CUBLAS_STATUS_ARCH_MISMATCH    // The function requires a feature absent from the device architecture; usually caused by the lack of support for double precision.
-	CublasStatusMappingError   Status = C.CUBLAS_STATUS_MAPPING_ERROR    // An access to GPU memory space failed, which is usually caused by a failure to bind a texture.
-	CublasStatusExecFailed     Status = C.CUBLAS_STATUS_EXECUTION_FAILED // The GPU program failed to execute. This is often caused by a launch failure of the kernel on the GPU, which can be caused by multiple reasons.
-	CublasStatusInternalError  Status = C.CUBLAS_STATUS_INTERNAL_ERROR   // An internal cuBLAS operation failed. This error is usually caused by a cudaMemcpyAsync() failure.
-	CublasStatusUnsupported    Status = C.CUBLAS_STATUS_NOT_SUPPORTED    // The functionnality requested is not supported
-	CublasStatusLicenceError   Status = C.CUBLAS_STATUS_LICENSE_ERROR    // The functionnality requested requires some license and an error was detected when trying to check the current licensing.
+	CublasStatusSuccess        Status = 0  // The operation completed successfully.
+	CublasStatusNotInitialized Status = 1  // The cuBLAS library was not initialized.
+	CublasStatusAllocFailed    Status = 3  // Resource allocation failed inside the cuBLAS library.
+	CublasStatusInvalidValue   Status = 7  // An unsupported value or parameter was passed to the function.
+	CublasStatusArchMismatch   Status = 8  // The function requires a feature absent from the device architecture.
+	CublasStatusMappingError   Status = 11 // An access to GPU memory space failed.
+	CublasStatusExecFailed     Status = 13 // The GPU program failed to execute.
+	CublasStatusInternalError  Status = 14 // An internal cuBLAS operation failed.
+	CublasStatusUnsupported    Status = 15 // The functionality requested is not supported.
+	CublasStatusLicenceError   Status = 16 // The functionality requested requires some license.
 )
 
 var StatusMessages = map[Status]string{
