+++ cublas/extension_cuda.go
@@ -1,0 +1,3 @@
+//go:build cuda || cuda11 || cuda12 || cuda13
+
+package cublas
