--- cublas/extension_stub.go
@@ -1,4 +1,0 @@
-//go:build !cuda
-// +build !cuda
-
-package cublas
